Converting between bytes is the responsibility of plugin-pb-go
